restapi/requirements.txt
Django==3.2
mysqlclient==2.1.0
django-mysql==4.5.0
djangorestframework==3.12.4
requests==2.26.0
drf_yasg==1.20.0
django_redis==5.0.0
